do not recreate the sage file list if it is present
Former-commit-id: fa73a40bceeed355183698f082512bc54ea0c9f8

+++ jacs2-services/src/main/java/org/janelia/jacs2/asyncservice/dataimport/SageLoaderProcessor.java
@@ -140,9 +140,11 @@ public class SageLoaderProcessor extends AbstractExeBasedServiceProcessor<Void, 
             Path workingDirectory = getWorkingDirectory(jacsServiceData);
             Files.createDirectories(workingDirectory);
             Path sageWorkingFile = FileUtils.getFilePath(workingDirectory, "SageFileList.txt");
-            Set<PosixFilePermission> perms = PosixFilePermissions.fromString("rwxrwx---");
-            Files.createFile(sageWorkingFile, PosixFilePermissions.asFileAttribute(perms)).toFile();
-            Files.write(sageWorkingFile, args.sampleFiles, StandardOpenOption.WRITE);
+            if (!Files.exists(sageWorkingFile)) {
+                Set<PosixFilePermission> perms = PosixFilePermissions.fromString("rw-rw----");
+                Files.createFile(sageWorkingFile, PosixFilePermissions.asFileAttribute(perms)).toFile();
+                Files.write(sageWorkingFile, args.sampleFiles, StandardOpenOption.WRITE);
+            }
             return sageWorkingFile;
         } catch (IOException e) {
             throw new UncheckedIOException(e);
